refactor: store folder structure
